==> 04-Pandas/notebooks/02_series.ipynb <==
# Pandas Series

A Series is a one-dimensional labeled array.

It can store:<br>
numbers, strings, booleans, etc.

Think of Series as:<br>
a single column of a table.

import pandas as pd

## Creating Series from List

data = [10,20,30,40]
s = pd.Series(data)

s

Each value has an index automatically assigned.

s.index
s.values

## Custom Index

s = pd.Series([100,200,300], index=["a","b","c"])
s

## Accessing Elements

s["a"]
s["c"]

## Mathematical Operations

s + 10
s * 2

## Series from Dictionary

data = {"Delhi":40, "Mumbai":35, "Chennai":38}
temp = pd.Series(data)

temp

## Handling Missing Values

temp.isnull()
temp.notnull()

## Basic Statistics

temp.mean()
temp.max()
temp.min()

## Conclusion

Series is the building block of DataFrame.
