==> 04-Pandas/notebooks/03_dataframe.ipynb <==
# Pandas DataFrame

A DataFrame is a two-dimensional labeled data structure.

It contains:<br>
rows + columns

Each column is a Pandas Series.

import pandas as pd

## Creating DataFrame from Dictionary

data = {
    "Name": ["Lakshay", "Amit", "Harsh"],
    "Age": [23, 25, 22],
    "City": ["Delhi", "Mumbai", "Pune"]
}

df = pd.DataFrame(data)
df

## Columns & Index

df.columns
df.index

## Selecting a Column (Series)

df["Name"]

## Selecting Multiple Columns

df[["Name", "City"]]

## Adding New Column

df["Salary"] = [50000, 60000, 55000]
df

## Deleting Column

df.drop("Salary", axis=1)

## Selecting Rows using loc

df.loc[0]
df.loc[1]

## Selecting Rows using iloc

df.iloc[0]
df.iloc[1]

## Basic Info

df.info()
df.describe()

## Conclusion

DataFrame is the main structure used for dataset manipulation.
