==> 04-Pandas/notebooks/05_inspection_selection.ipynb <==
# Inspection, Selection & Filtering

After loading data, we explore and select useful parts of the dataset.

import pandas as pd

sales = pd.read_csv("../data/raw/sales.csv")
sales.head()

## Selecting Columns

sales["product"]

sales[["product", "price"]]

## Selecting Rows

sales.loc[0]
sales.loc[2]

sales.iloc[0]
sales.iloc[1]

## Conditional Filtering

sales[sales["price"] > 5000]

sales[(sales["city"] == "Delhi") & (sales["price"] > 2000)]

## Using Query Method

sales.query("price > 5000")

sales.query("city == 'Delhi'")

## Sorting Data

sales.sort_values("price")

sales.sort_values("price", ascending=False)

## Conclusion

Filtering and selecting data is a core step before cleaning and analysis.
